feat : question, response 코드 분리
